Part 1: Remove legacy api
