+++ src/olympia/amo/tests/test_middleware.py
@@ -37,7 +37,6 @@ class TestMiddleware(TestCase):
     def test_authentication_used_outside_the_api(self, process_request):
         req = RequestFactory().get('/')
         req.is_api = False
-        req.is_legacy_api = False
         AuthenticationMiddlewareWithoutAPI().process_request(req)
         assert process_request.called
 
@@ -46,13 +45,11 @@ class TestMiddleware(TestCase):
     def test_authentication_not_used_with_the_api(self, process_request):
         req = RequestFactory().get('/')
         req.is_api = True
-        req.is_legacy_api = False
         AuthenticationMiddlewareWithoutAPI().process_request(req)
         assert not process_request.called
 
         req = RequestFactory().get('/')
         req.is_api = False
-        req.is_legacy_api = True
         AuthenticationMiddlewareWithoutAPI().process_request(req)
         assert not process_request.called
