+++ src/olympia/amo/tests/test_url_prefix.py
@@ -81,18 +81,6 @@ class MiddlewareTest(TestCase):
         response = self.process('/api/v4/some/endpoint/')
         assert response is None
 
-    def test_v1_api_is_identified_as_api_request(self):
-        response = self.process('/en-US/firefox/api/')
-        assert response is None
-        assert self.request.LANG == 'en-US'
-        assert self.request.is_legacy_api
-
-        # double-check _only_ /api/ is marked as .is_api
-        response = self.process('/en-US/firefox/apii/')
-        assert response is None
-        assert self.request.LANG == 'en-US'
-        assert not self.request.is_legacy_api
-
     def test_vary(self):
         response = self.process('/')
         assert response['Vary'] == 'Accept-Language, User-Agent'
