--- src/olympia/legacy_api/__init__.py
@@ -1,3 +1,0 @@
-MIN_VERSION = 1.0
-CURRENT_VERSION = 1.5
-MAX_VERSION = CURRENT_VERSION
