--- src/olympia/legacy_api/templates/legacy_api/addon_detail.xml
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-{% include 'legacy_api/includes/addon.xml' %}
