--- src/olympia/legacy_api/templates/legacy_api/compat.xml
@@ -1,27 +1,0 @@
-{% for c in compat %}
-<addon_compatibility
-    hosted="{{ c.is_hosted()|json }}"
-    {% if c.addon_id %}id="{{ c.addon_id }}"{% endif %}>
-    <guid>{{ c.guid }}</guid>
-    <name>{{ c.name }}</name>
-    <version_ranges>
-      {% for vr in c.collapsed_ranges() %}
-        <version_range type="{{ vr.type }}">
-          <min_version>{{ vr.min }}</min_version>
-          <max_version>{{ vr.max }}</max_version>
-          <compatible_applications>
-            {% for app in vr.apps %}
-              <application>
-                <name>{{ app.app.pretty }}</name>
-                <application_id>{{ app.app.id }}</application_id>
-                <min_version>{{ app.min }}</min_version>
-                <max_version>{{ app.max }}</max_version>
-                <appID>{{ app.app.guid }}</appID>
-              </application>
-            {% endfor %}
-          </compatible_applications>
-        </version_range>
-      {% endfor %}
-    </version_ranges>
-</addon_compatibility>
-{% endfor %}
