--- src/olympia/legacy_api/templates/legacy_api/includes/addon.xml
@@ -1,130 +1,0 @@
-{% set new_api = (api_version >= 1.5) %}
-{% if addon.compat_version %}
-  {% set version = addon.compat_version %}
-{% else %}
-  {% set version = addon.current_version %}
-{% endif %}
-<addon id="{{ addon.id }}">
-  <name>{{ addon.name }}</name>
-  <type id="{{ addon.type }}">{{ amo.ADDON_TYPE[addon.type] }}</type>
-  <guid>{{ addon.guid }}</guid>
-  <slug>{{ addon.slug }}</slug>
-  <version>{{ version.version }}</version>
-  <status id="{{ addon.status }}">{{ addon.STATUS_CHOICES[addon.status] }}</status>
-  <authors>
-    {% if not addon.is_persona() %}
-    {% for author in addon.listed_authors -%}
-    {%- if new_api -%}
-    <author id="{{ author.id }}">
-      <name>{{ author.name }}</name>
-      <link>{{ author.get_url_path()|absolutify|urlparams(src='api') }}</link>
-    </author>
-    {%- else -%}
-    <author>{{ author.name }}</author>
-    {%- endif -%}
-    {%- endfor %}
-    {% endif %}
-  </authors>
-  <summary>{%- if new_api -%}{{ addon.summary }}{% else %}{{ addon.summary|clean(true) }}{% endif -%}</summary>
-  <description>{%- if new_api %}{{ addon.description }}{% else -%}{{ addon.description|clean(true) }}{% endif %}</description>
-  {%- if version and version.license -%}
-  <license>
-    <name>{{ version.license.name }}</name>
-    <url>
-      {%- if not version.license.url -%}
-      {{ version.license_url()|absolutify }}
-      {%- else -%}
-      {{ version.license.url|absolutify }}
-      {%- endif -%}
-    </url>
-  </license>
-  {%- endif -%}
-  {# The 32px icon must be the last icon specified. #}
-  <icon size="64">{{ addon.get_icon_url(64, use_default=False) }}</icon>
-  <icon size="32">{{ addon.get_icon_url(32, use_default=False) }}</icon>
-  <compatible_applications>
-    {%- if version -%}
-    {%- for app in version.compatible_apps.values() %}
-    {%- if amo.APP_IDS.get(app.application) -%}
-    <application>
-      <name>{{ amo.APP_IDS[app.application].pretty }}</name>
-      <application_id>{{ app.application }}</application_id>
-      <min_version>{{ app.min }}</min_version>
-      <max_version>{{ app.max }}</max_version>
-      <appID>{{ amo.APP_IDS[app.application].guid }}</appID>
-    </application>
-    {%- endif -%}
-    {%- endfor -%}
-    {%- endif -%}
-  </compatible_applications>
-  <all_compatible_os>
-    {%- if version -%}
-    {% for os in version.supported_platforms -%}
-    <os>{{ os.api_name }}</os>
-    {%- endfor -%}
-    {%- endif -%}
-  </all_compatible_os>
-  <eula>{%- if new_api %}{{ addon.eula }}{% else %}{{ addon.eula|clean(true) }}{% endif -%}</eula>
-  {%- if new_api -%}
-  <previews>
-    {%- for preview in addon.current_previews -%}
-    <preview position="{{ preview.position|int }}">
-      <full type="image/png"{% with sizes=preview.image_size -%}{% if sizes %} width="{{ sizes[0] }}" height="{{ sizes[1] }}"{% endif %}{% endwith -%}>
-        {{ preview.image_url|urlparams(src='api') }}
-      </full>
-      <thumbnail type="image/png"{% with sizes=preview.thumbnail_size -%}{% if sizes %} width="{{ sizes[0] }}" height="{{ sizes[1] }}"{% endif %}{% endwith -%}>
-        {{ preview.thumbnail_url|urlparams(src='api') }}
-      </thumbnail>
-      {%- if preview.caption -%}
-        <caption>{{ preview.caption }}</caption>
-      {%- endif -%}
-    </preview>
-    {%- endfor -%}
-  </previews>
-  {%- else -%}
-  <thumbnail>{{ addon.thumbnail_url }}</thumbnail>
-  {%- endif -%}
-  <rating>{{ addon.average_rating|round|int }}</rating>
-  <learnmore>{{ addon.get_url_path()|absolutify|urlparams(src='api') }}</learnmore>
-  {%- if version -%}
-  {%- for file in version.all_files -%}
-  <install hash="{{ file.hash }}"
-    os="{{ amo.PLATFORMS[file.platform].api_name }}"
-    size="{{ file.size }}">{{ file.get_url_path('api') }}</install>
-  {% endfor -%}
-  {%- endif -%}
-  {%- if new_api %}
-    {%- if addon.contributions -%}
-    <contribution_data>
-      <meet_developers>
-        {{ addon.contributions }}
-      </meet_developers>
-    </contribution_data>
-    {%- endif -%}
-    <developer_comments>{{ addon.developer_comments }}</developer_comments>
-    <reviews num="{{ addon.total_ratings }}">
-      {{ addon.ratings_url|absolutify|urlparams(src='api') }}
-    </reviews>
-    <total_downloads>{{ addon.total_downloads }}</total_downloads>
-    <weekly_downloads>{{ addon.weekly_downloads }}</weekly_downloads>
-    <daily_users>{{ addon.average_daily_users }}</daily_users>
-    <created epoch="{{ addon.created|epoch }}">
-      {{ addon.created|isotime }}
-    </created>
-    <last_updated epoch="{{ addon.last_updated|epoch }}">
-      {{ addon.last_updated|isotime }}
-    </last_updated>
-    <homepage>{{ addon.homepage }}</homepage>
-    <support>{{ addon.support_url }}</support>
-    <featured>{{ addon.is_featured(request.APP, request.LANG)|int }}</featured>
-    {# Only kept for backwards compatibility of the API #}
-    <performance>
-    </performance>
-
-    {% if show_localepicker -%}
-    <strings><![CDATA[
-{{ addon.get_localepicker() }}
-    ]]></strings>
-    {% endif %}
-  {%- endif -%}
-</addon>
