--- src/olympia/legacy_api/templates/legacy_api/list.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<addons>
-  {% for addon in addons %}
-    {% include 'legacy_api/includes/addon.xml' %}
-  {% endfor %}
-</addons>
