--- src/olympia/legacy_api/templates/legacy_api/message.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-{# Compatibility across both Piston and DRF on error messages/vars. #}
-{% if detail %}
-  <error>{{_(detail)}}</error>
-{% else %}
-  {% if error_level %}
-    <error>{{_(msg)}}</error>
-  {% else %}
-    <msg>{{_(msg)}}</msg>
-  {% endif %}
-{% endif %}
