--- src/olympia/legacy_api/templates/legacy_api/search.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<searchresults total_results="{{ total }}">
-  {% if addons_xml %}
-      {% for xml in addons_xml %}
-          {{ xml|xssafe }}
-      {% endfor %}
-  {% else %}
-      {% for addon in results -%}
-          {% include 'legacy_api/includes/addon.xml' -%}
-      {% endfor %}
-  {% endif %}
-  {% if compat %}{% include 'legacy_api/compat.xml' %}{% endif %}
-</searchresults>
